+++ backend/src/controllers/post.js
@@ -43,12 +43,22 @@ exports.modifyPost = (req, res, next) => {
 // Supprimer un post
 exports.deletePost = (req, res, next) => {
     const id = req.params.id
+    const idUser = req.id_user
+    let post;
 
-    db.query(`DELETE FROM post WHERE id = ?`, id, (error, result) => {
-        if (error) {
-            return res.status(400).json({ error: "Le post n'a pas pu être supprimé" })
+    db.query(`SELECT * FROM post WHERE id = ?`, id, (error, result) => {
+        post = result[0]
+
+        if(post.id_user === idUser || req.admin === 1) {
+            db.query(`DELETE FROM post WHERE id = ?`, id, (error, result) => {
+                if (error) {
+                    return res.status(400).json({ error: "Le post n'a pas pu être supprimé" })
+                }
+                return res.status(200).json(result);
+            })
+        } else {
+            return res.status(403).json({ error: "Vous ne pouvez pas supprimer le post !"})
         }
-        return res.status(200).json(result);
     })
 }
